Extract shared Koa bootstrap into src/app-common.js
app-www.js and app-download.js duplicated ~150 lines of identical server
setup. Move the shared pieces into app-common.js:
  - createBaseApp(): Koa app + logger + INI config + GeoDb + MySQL
  - useObservability(): response-time, access/error logging, Prometheus
  - useTimeout(): 8s request timeout
  - useCompression(): gzip/brotli/zstd (brotli quality and zstd level
    differ per server, so they are parameters)
  - useResponseLength(): record response length for 200s
  - startServer(): PID file, SIGHUP handling, listen

Each entry point now keeps only its domain-specific middleware. The
import.meta.url run-directly guard stays in the entry points so the
server only starts when run directly.
